+++ tests/compilers/tag-if-control-statements.test.js
@@ -259,4 +259,63 @@ describe('oval-control-statements if', function () {
     var compiled = compiler.compile(content)
     expect(compiled.trim().replace(/ /g, '').replace(/\n/g, '')).to.eq(expectedCompiledCode.trim().replace(/ /g, '').replace(/\n/g, ''))
   })
+
+  it('compiles a statement with nested statements', function () {
+    var content = `
+      <tag-name>
+        <div if={tag.value > 0}
+          class="test-class">
+          <div if={tag.value < 0}>
+            text
+          </div>
+        </div>
+      </tag-name>
+    `
+
+    var expectedCompiledCode = `
+      <tag-name>
+        {
+          tag.value > 0
+            ? (
+              <div class="test-class">
+                {
+                  tag.value < 0 ? (
+                    <div>
+                      text
+                    </div>
+                  ) : null
+                }
+              </div>
+            )
+            : null
+        }
+      </tag-name>
+    `
+
+    var compiled = compiler.compile(content)
+    expect(compiled.trim().replace(/ /g, '').replace(/\n/g, '')).to.eq(expectedCompiledCode.trim().replace(/ /g, '').replace(/\n/g, ''))
+  })
+
+  it('compiles a single line statement with inner expressions', function () {
+    var content = `
+      <tag-name>
+        <span if={method(arg1, arg2)}>{fn1('arg1')} {fn2.inner(arg2)}</span>
+      </tag-name>
+    `
+
+    var expectedCompiledCode = `
+      <tag-name>
+        {
+          method(arg1, arg2)
+            ? (
+              <span>{fn1('arg1')} {fn2.inner(arg2)}</span>
+            )
+            : null
+        }
+      </tag-name>
+    `
+
+    var compiled = compiler.compile(content)
+    expect(compiled.trim().replace(/ /g, '').replace(/\n/g, '')).to.eq(expectedCompiledCode.trim().replace(/ /g, '').replace(/\n/g, ''))
+  })
 })